ActuatorEffectivenessRotors: avoid early return on parameter load failure
This lead to a compiler warning with arm-none-eabi-gcc 12.2.0
about the variable count potentially being a dangling pointer.

Since it's an early return the case is not possible but it's nicer to
have no early return at all and I didn't want to disable the
copile warning explicitly for this case.